OGC API Part 4 implementation

+++ OgcApi.Net.Features/DataProviders/IDataProvider.cs
@@ -14,7 +14,7 @@ namespace OgcApi.Net.Features.DataProviders
 
         OgcFeature GetFeature(string collectionId, string featureId, string apiKey = null);
 
-        int GetFeaturesCount(string collectionId, Envelope bbox = null, DateTime? startDateTime = null, DateTime? endDateTime = null, string apiKey = null);
+        long GetFeaturesCount(string collectionId, Envelope bbox = null, DateTime? startDateTime = null, DateTime? endDateTime = null, string apiKey = null);
 
         string CreateFeature(string collectionId, OgcFeature feature, string apiKey = null);
